Update README.md
Sentiment of tweets are analyzed as positive, negative and neutral. Tweepy is used to collect tweets which requires twitter development account. Tweets are cleaned as they contain RT,url @username etc. Also stopwords, punctuation are removed as part of data cleaning. TfidfVectorizer is used for feature extraction. Machine Learning model used is Random Forest. Accuracy for this model is 84%.